requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.2
idna==2.10
isort==5.8.0
kombu==5.0.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
rabbitmq==0.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
